+++ PeepPeep/DeviceActivityReportExtension/Model/EmptyModel.swift
@@ -1,0 +1,8 @@
+//
+//  EmptyModel.swift
+//  DevuceActuvityReportExtension
+//
+//  Created by Ha Jong Myeong on 2023/07/17.
+//
+
+import Foundation
