+++ PeepPeep/DeviceActivityReportExtension/View/CostumeBook/CostumeBookView.swift
@@ -1,0 +1,20 @@
+//
+//  CostumeBookView.swift
+//  DevuceActuvityReportExtension
+//
+//  Created by Ha Jong Myeong on 2023/07/18.
+//
+
+import SwiftUI
+
+struct CostumeBookView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct CostumeBookView_Previews: PreviewProvider {
+    static var previews: some View {
+        CostumeBookView()
+    }
+}
