+++ PeepPeep/DeviceActivityReportExtension/View/FinalReport/FinalReportView.swift
@@ -1,0 +1,20 @@
+//
+//  FinalReportView.swift
+//  DevuceActuvityReportExtension
+//
+//  Created by Ha Jong Myeong on 2023/07/18.
+//
+
+import SwiftUI
+
+struct FinalReportView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct FinalReportView_Previews: PreviewProvider {
+    static var previews: some View {
+        FinalReportView()
+    }
+}
