+++ PeepPeep/DeviceActivityReportExtension/View/MainScreen/MainView.swift
@@ -1,0 +1,20 @@
+//
+//  MainView.swift
+//  DevuceActuvityReportExtension
+//
+//  Created by Ha Jong Myeong on 2023/07/18.
+//
+
+import SwiftUI
+
+struct MainView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct MainView_Previews: PreviewProvider {
+    static var previews: some View {
+        MainView()
+    }
+}
