+++ PeepPeep/DeviceActivityReportExtension/View/SplashScreen/SplashView.swift
@@ -1,0 +1,20 @@
+//
+//  SplashView.swift
+//  DevuceActuvityReportExtension
+//
+//  Created by Ha Jong Myeong on 2023/07/18.
+//
+
+import SwiftUI
+
+struct SplashView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct SplashView_Previews: PreviewProvider {
+    static var previews: some View {
+        SplashView()
+    }
+}
