+++ PeepPeep/DeviceActivityReportExtension/ViewModel/EmptyViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  EmptyViewModel.swift
+//  DevuceActivityReportExtension
+//
+//  Created by Ha Jong Myeong on 2023/07/17.
+//
+
+import Foundation
