--- PeepPeep/FeedBack/Common/Components/EmptyCompnents.swift
@@ -1,8 +1,0 @@
-//
-//  EmptyCompnents.swift
-//  FeedBack
-//
-//  Created by Ha Jong Myeong on 2023/07/17.
-//
-
-import Foundation
