--- PeepPeep/FeedBack/Common/Extensions/EmptyExtensions.swift
@@ -1,8 +1,0 @@
-//
-//  EmptyExtensions.swift
-//  FeedBack
-//
-//  Created by Ha Jong Myeong on 2023/07/17.
-//
-
-import Foundation
