--- PeepPeep/FeedBack/Common/Themes/Colors.swift
@@ -1,8 +1,0 @@
-//
-//  Colors.swift
-//  FeedBack
-//
-//  Created by Ha Jong Myeong on 2023/07/17.
-//
-
-import Foundation
