--- PeepPeep/FeedBack/Common/Themes/Styles.swift
@@ -1,8 +1,0 @@
-//
-//  Styles.swift
-//  FeedBack
-//
-//  Created by Ha Jong Myeong on 2023/07/17.
-//
-
-import Foundation
