--- PeepPeep/FeedBack/Common/Utils/EmptyUtils.swift
@@ -1,8 +1,0 @@
-//
-//  EmptyUtils.swift
-//  FeedBack
-//
-//  Created by Ha Jong Myeong on 2023/07/17.
-//
-
-import Foundation
