--- PeepPeep/FeedBack/Model/EmptyModel.swift
@@ -1,8 +1,0 @@
-//
-//  EmptyModel.swift
-//  FeedBack
-//
-//  Created by Ha Jong Myeong on 2023/07/17.
-//
-
-import Foundation
