--- PeepPeep/FeedBack/View/CostumeBook/CostumeBookView.swift
@@ -1,20 +1,0 @@
-//
-//  CostumeBookView.swift
-//  FeedBack
-//
-//  Created by Ha Jong Myeong on 2023/07/18.
-//
-
-import SwiftUI
-
-struct CostumeBookView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct CostumeBookView_Previews: PreviewProvider {
-    static var previews: some View {
-        CostumeBookView()
-    }
-}
