--- PeepPeep/FeedBack/View/DailyFeedBack/DailyFeedBackView.swift
@@ -1,20 +1,0 @@
-//
-//  DailyFeedBackView.swift
-//  FeedBack
-//
-//  Created by Ha Jong Myeong on 2023/07/18.
-//
-
-import SwiftUI
-
-struct DailyFeedBackView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct DailyFeedBackView_Previews: PreviewProvider {
-    static var previews: some View {
-        DailyFeedBackView()
-    }
-}
