--- PeepPeep/FeedBack/View/MainScreen/MainView.swift
@@ -1,20 +1,0 @@
-//
-//  MainView.swift
-//  FeedBack
-//
-//  Created by Ha Jong Myeong on 2023/07/18.
-//
-
-import SwiftUI
-
-struct MainView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct MainView_Previews: PreviewProvider {
-    static var previews: some View {
-        MainView()
-    }
-}
