--- PeepPeep/FeedBack/View/WeeklyFeedBack/WeeklyFeedBackView.swift
@@ -1,20 +1,0 @@
-//
-//  WeeklyFeedBackView.swift
-//  FeedBack
-//
-//  Created by Ha Jong Myeong on 2023/07/18.
-//
-
-import SwiftUI
-
-struct WeeklyFeedBackView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct WeeklyFeedBackView_Previews: PreviewProvider {
-    static var previews: some View {
-        WeeklyFeedBackView()
-    }
-}
