--- PeepPeep/FeedBack/ViewModel/EmptyViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  EmptyViewModel.swift
-//  FeedBack
-//
-//  Created by Ha Jong Myeong on 2023/07/17.
-//
-
-import Foundation
