+++ PeepPeep/PeepPeep.xcodeproj/xcshareddata/xcschemes/PeepPeep.xcscheme
@@ -28,6 +28,19 @@
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
       shouldUseLaunchSchemeArgsEnv = "YES"
       shouldAutocreateTestPlan = "YES">
+      <Testables>
+         <TestableReference
+            skipped = "NO"
+            parallelizable = "YES">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "C910C9BD2A67D8A300EEF718"
+               BuildableName = "PeepPeepCommonsTests.xctest"
+               BlueprintName = "PeepPeepCommonsTests"
+               ReferencedContainer = "container:PeepPeep.xcodeproj">
+            </BuildableReference>
+         </TestableReference>
+      </Testables>
    </TestAction>
    <LaunchAction
       buildConfiguration = "Debug"
