+++ PeepPeep/PeepPeepCommons/PeepPeepCommons.h
@@ -1,0 +1,18 @@
+//
+//  PeepPeepCommons.h
+//  PeepPeepCommons
+//
+//  Created by Ha Jong Myeong on 2023/07/19.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for PeepPeepCommons.
+FOUNDATION_EXPORT double PeepPeepCommonsVersionNumber;
+
+//! Project version string for PeepPeepCommons.
+FOUNDATION_EXPORT const unsigned char PeepPeepCommonsVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <PeepPeepCommons/PublicHeader.h>
+
+
