+++ PeepPeep/PeepPeepCommons/Utils/EmptyUtils.swift
